Consistently use int64 for integers holding number of milliseconds.
This applies only to things which use TimeDelta::InMilliseconds, because
it returns int64. Before this patch callers static_casted that to int,
and after this patch they use the returned value as int64.

Review URL: http://codereview.chromium.org/126279

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@18626 0039d316-1c4b-4281-b951-d872f2087c98

+++ base/message_loop.cc
@@ -232,7 +232,7 @@ void MessageLoop::PostTask(
 }
 
 void MessageLoop::PostDelayedTask(
-    const tracked_objects::Location& from_here, Task* task, int delay_ms) {
+    const tracked_objects::Location& from_here, Task* task, int64 delay_ms) {
   PostTask_Helper(from_here, task, delay_ms, true);
 }
 
@@ -242,13 +242,13 @@ void MessageLoop::PostNonNestableTask(
 }
 
 void MessageLoop::PostNonNestableDelayedTask(
-    const tracked_objects::Location& from_here, Task* task, int delay_ms) {
+    const tracked_objects::Location& from_here, Task* task, int64 delay_ms) {
   PostTask_Helper(from_here, task, delay_ms, false);
 }
 
 // Possibly called on a background thread!
 void MessageLoop::PostTask_Helper(
-    const tracked_objects::Location& from_here, Task* task, int delay_ms,
+    const tracked_objects::Location& from_here, Task* task, int64 delay_ms,
     bool nestable) {
   task->SetBirthPlace(from_here);
